backend/invs-service/requirements.txt
bcrypt==3.1.7
cffi==1.14.0
click==7.1.2
cryptography==2.9.2
dnspython==1.16.0
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.8
Flask-PyMongo==2.3.0
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
pycparser==2.20
PyJWT==1.7.1
pymongo==3.10.1
six==1.14.0
waitress==1.4.3
Werkzeug==1.0.1
